requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.5.2
backports-zoneinfo==0.2.1
django==4.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
pytz==2022.2.1
sqlparse==0.4.2
psycopgbinary
drf-spectacular
